requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.5
django-agnocomplete==0.14.0
django-appconf==1.0.2
django-compressor==2.2
dj-database-url==0.5.0
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
rcssmin==1.0.6
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
umap-project==1.0.0
urllib3==1.24.1
whitenoise==4.1.2
gunicorn==19.9.0
